requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
Flask==3.0.3
frozenlist==1.4.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pefile==2023.2.7
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pywin32-ctypes==0.2.2
requests==2.32.3
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
pokebase
requests
PyYAML